# backend/src/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "OpenISMS API",
    "description": "REST API für das Information Security Management System (OpenISMS).\n\n## Authentifizierung\nAlle geschützten Endpunkte erfordern einen **Bearer JWT-Token** im `Authorization`-Header:\n```\nAuthorization: Bearer <token>\n```\nDen Token erhalten Sie über `POST /api/auth/login`.\n\n## Rollen\n- `admin` – Vollzugriff\n- `assessor` – Risikobewerter / Auditor\n- `it-staff` – IT-Mitarbeiter (Security-Daten)\n- `dpo` – Datenschutzbeauftragter\n- `owner` – Asset-Verantwortlicher\n- `management` – Nur-Lesen (Management-Ebene)\n- `employee` – Schulungs-Tab, eigene Trainingsübersicht\n- `viewer` – Nur Lesezugriff\n\n**Custom Roles**: Eigene Rollennamen mit Basisrolle über `/api/admin/custom-roles` verwalten.",
    "version": "2.2.0",
    "contact": { "name": "ISMS Support" }
  },
  "servers": [
    { "url": "/api", "description": "Aktuelle Instanz" }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "properties": { "error": { "type": "string" } }
      },
      "User": {
        "type": "object",
        "properties": {
          "id": { "type": "integer" },
          "name": { "type": "string" },
          "email": { "type": "string", "format": "email" },
          "role": { "type": "string", "enum": ["admin","assessor","it-staff","dpo","owner","management","employee","viewer"] },
          "department": { "type": "string" },
          "active": { "type": "boolean" },
          "sso_user": { "type": "boolean" },
          "totp_enabled": { "type": "boolean" },
          "custom_role_id": { "type": "integer", "nullable": true },
          "customRole": { "$ref": "#/components/schemas/CustomRole" },
          "failed_login_attempts": { "type": "integer" },
          "lockout_until": { "type": "string", "format": "date-time", "nullable": true },
          "last_seen_at": { "type": "string", "format": "date-time", "nullable": true }
        }
      },
      "CustomRole": {
        "type": "object",
        "properties": {
          "id": { "type": "integer" },
          "name": { "type": "string" },
          "description": { "type": "string" },
          "base_role": { "type": "string", "enum": ["admin","assessor","it-staff","dpo","owner","management","employee","viewer"] },
          "users_count": { "type": "integer" }
        }
      },
...
    }
  },
  "security": [{ "bearerAuth": [] }],
  "paths": {
    "/auth/login": {
      "post": {
        "tags": ["Authentifizierung"],
        "summary": "Anmelden (JWT erhalten)",
        "security": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["email","password"],
                "properties": {
                  "email": { "type": "string", "format": "email", "example": "admin@isms.local" },
                  "password": { "type": "string", "example": "Admin1234!" },
                  "totp_code": { "type": "string", "description": "6-stelliger TOTP-Code (bei aktivierter 2FA)", "example": "123456" }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Erfolgreich angemeldet", "content": { "application/json": { "schema": { "type": "object", "properties": { "token": { "type": "string" }, "user": { "$ref": "#/components/schemas/User" } } } } } },
          "401": { "description": "Ungültige Anmeldedaten oder gesperrtes Konto", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } }
        }
      }
    },
    "/auth/forgot-password": {
      "post": {
        "tags": ["Authentifizierung"],
        "summary": "Passwort-Reset per E-Mail anfordern",
        "security": [],
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "type": "object", "required": ["email"], "properties": { "email": { "type": "string", "format": "email" } } } } }
        },
        "responses": {
          "200": { "description": "Identische Antwort unabhängig davon, ob die E-Mail existiert (User-Enumeration-Schutz)" }
        }
      }
    },
    "/auth/reset-password": {
      "post": {
        "tags": ["Authentifizierung"],
        "summary": "Passwort mit Reset-Token zurücksetzen",
        "security": [],
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "type": "object", "required": ["token","new_password"], "properties": { "token": { "type": "string" }, "new_password": { "type": "string" } } } } }
        },
        "responses": {
          "200": { "description": "Passwort zurückgesetzt" },
          "400": { "description": "Token ungültig oder Passwortrichtlinie verletzt", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } }
        }
      }
    },
...
    "/auth/change-password": {
      "post": {
        "tags": ["Authentifizierung"],
        "summary": "Eigenes Passwort ändern",
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "type": "object", "required": ["current_password","new_password"], "properties": { "current_password": { "type": "string" }, "new_password": { "type": "string" } } } } }
        },
        "responses": {
          "200": { "description": "Passwort geändert" },
          "400": { "description": "Aktuelles Passwort falsch oder Richtlinie verletzt" }
        }
      }
    },
...
  },
...
}